pyproject.toml
[tool.poetry]
name = "repl_python3_Goblin-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
replit = "^3.2.4"
Flask = "^2.0.2"
bs4 = "^0.0.1"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
